Moved WhoisServiceUtils methods into RestServiceHelper.
